refactor: organiza view e arquivos user admin
