+++ src/services/User/UserService.js
@@ -15,6 +15,16 @@ class UserService {
     const response = await api.put(`users/${id}`, body, config)
     return response.data
   }
+
+  async getAllUsers() {
+    const response = await api.get('users')
+    return response
+  }
+
+  async updateStatusUserDelete(userId, body) {
+      const response = await api.put(`users/${userId}`, body)
+      return response
+  }
 }
 
 export default new UserService()
